requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.51.0
pandas==2.2.3
openpyxl==3.1.5
httpx==0.27.2
python-dotenv==1.0.1
apscheduler==3.10.4
pydantic==2.9.2
